requirements.txt
okx-sdk
pandas
ollama
